Rename cctools -> scratch-utils; address review feedback
Rename throughout: cctools/ -> scratch-utils/, cmd/cctools -> cmd/scratch-utils,
images/calico-cctools -> images/calico-scratch-utils, calico/cctools ->
calico/scratch-utils, and the CALICO_CCTOOLS make vars. The ccutil package moves
with it as scratch-utils/util.

Review fixes:

- go.mod: drop golang.org/x/tools and honnef.co/go/tools. Nothing here imports
  them; they leaked in from an unrelated local experiment and dragged a large
  tree into go.sum. Re-tidied.

- gce.PutData: errBuf was declared but never wired to sess.Stderr, so every
  upload failure reported an empty remote stderr. Wire it up, and do the same
  for PutDir, which was equally silent.

- createvm/runonvm: bound the compute-API calls with a context deadline.
  waitZoneOp loops on ZoneOperations.Wait, so an operation that never reaches
  DONE hung the step until the workflow's own timeout. deletevm already had one.
  runonvm scopes its deadline to setup only -- running the script is the CI job
  and stays unbounded.

- cmd/Makefile: bin/scratch-utils-$(ARCH) depended only on main.go, so editing
  anything under scratch-utils/ left an incremental build shipping a stale
  binary. Depend on the whole source tree.

- Drop stale comments: there is no WaitReady, guest attributes are never
  enabled, and the cloud/kind-rig/... paths are from the origin repo.
