+++ .gitignore
@@ -2,4 +2,4 @@
 cmd/bin
 images/calico-binfmt/bin
 images/calico-go-build/bin
-images/calico-cctools/bin
+images/calico-scratch-utils/bin
