+++ .semaphore/promotions/calico-scratch-utils.yml
@@ -1,5 +1,5 @@
 version: v1.0
-name: Publish calico/cctools images
+name: Publish calico/scratch-utils images
 agent:
   machine:
     type: f1-standard-2
@@ -20,7 +20,7 @@ global_job_config:
       - checkout
 
 blocks:
-  - name: Publish calico/cctools amd64 images
+  - name: Publish calico/scratch-utils amd64 images
     dependencies: []
     run:
       when: "branch = 'master'"
@@ -28,4 +28,4 @@ blocks:
       jobs:
         - name: Linux amd64
           commands:
-            - if [ -z "${SEMAPHORE_GIT_PR_NUMBER}" ]; then make -C images calico-cctools-cd CONFIRM=true; fi
+            - if [ -z "${SEMAPHORE_GIT_PR_NUMBER}" ]; then make -C images calico-scratch-utils-cd CONFIRM=true; fi
