+++ .semaphore/semaphore.yml
@@ -48,11 +48,19 @@ promotions:
     pipeline_file: promotions/calico-tinygo.yml
     auto_promote:
       when: "branch = 'master' AND change_in('/images/calico-tinygo/')"
-  # Publish cctools images for master when the binary source or image changed.
-  - name: Publish calico/cctools images
-    pipeline_file: promotions/calico-cctools.yml
+  # Publish scratch-utils images for master when the binary source or image changed.
+  - name: Publish calico/scratch-utils images
+    pipeline_file: promotions/calico-scratch-utils.yml
     auto_promote:
-      when: "branch = 'master' AND change_in(['/images/calico-cctools/', '/cmd/cctools/', '/cctools/'])"
+      when: "branch = 'master' AND change_in(['/images/calico-scratch-utils/', '/cmd/scratch-utils/', '/scratch-utils/'])"
+  # Manual only (no auto_promote): trigger from the Semaphore UI to (re)build the
+  # ci-base GCE VM image. See promotions/ci-base-vm-image.yml and vm-image/.
+  - name: Build ci-base VM image
+    pipeline_file: promotions/ci-base-vm-image.yml
+  # Manual only (no auto_promote): trigger from the Semaphore UI to (re)build the
+  # go-build preload disk image. See promotions/preload-disk-image.yml and preload-disk/.
+  - name: Build go-build preload disk image
+    pipeline_file: promotions/preload-disk-image.yml
 
 blocks:
   - name: calico/go-build image
@@ -192,10 +200,10 @@ blocks:
             - env_var: ARCH
               values: ["amd64", "arm64"]
 
-  - name: calico/cctools image
+  - name: calico/scratch-utils image
     dependencies: []
     task:
       jobs:
-        - name: Build calico/cctools amd64 image
+        - name: Build calico/scratch-utils amd64 image
           commands:
-            - make -C images calico-cctools-image ARCH=amd64
+            - make -C images calico-scratch-utils-image ARCH=amd64
