+++ images/Makefile
@@ -6,21 +6,21 @@ CALICO_BINFMT ?= binfmt
 CALICO_GO_BUILD ?= go-build
 CALICO_RUST_BUILD ?= rust-build
 CALICO_TINYGO ?= tinygo
-CALICO_CCTOOLS ?= cctools
+CALICO_SCRATCH_UTILS ?= scratch-utils
 
 .PHONY: build
 build:
 	$(MAKE) -C ../cmd build
-	mkdir -p calico-binfmt/bin calico-go-build/bin/ calico-cctools/bin/
+	mkdir -p calico-binfmt/bin calico-go-build/bin/ calico-scratch-utils/bin/
 	cp ../cmd/bin/binfmt-$(ARCH) calico-binfmt/bin/binfmt-$(ARCH)
 	cp ../cmd/bin/semvalidator-$(ARCH) calico-go-build/bin/semvalidator-$(ARCH)
-	cp ../cmd/bin/cctools-$(ARCH) calico-cctools/bin/cctools-$(ARCH)
+	cp ../cmd/bin/scratch-utils-$(ARCH) calico-scratch-utils/bin/scratch-utils-$(ARCH)
 
 .PHONY: image
-image: calico-base-image calico-binfmt-image calico-go-build-image calico-rust-build-image calico-tinygo-image calico-cctools-image
+image: calico-base-image calico-binfmt-image calico-go-build-image calico-rust-build-image calico-tinygo-image calico-scratch-utils-image
 
 .PHONY: image-all
-image-all: calico-base-image-all calico-binfmt-image calico-go-build-image-all calico-rust-build-image-all calico-tinygo-image-all calico-cctools-image
+image-all: calico-base-image-all calico-binfmt-image calico-go-build-image-all calico-rust-build-image-all calico-tinygo-image-all calico-scratch-utils-image
 
 ################################################################################
 # Base image for all Calico components.
@@ -106,20 +106,20 @@ calico-binfmt-cd: calico-binfmt-image var-require-one-of-CONFIRM-DRYRUN
 	$(MAKE) BUILD_IMAGES=$(CALICO_BINFMT) retag-build-images-with-registries push-images-to-registries IMAGETAG=$(CALICO_BINFMT_IMAGETAG) LATEST_IMAGE_TAG=$(CALICO_BINFMT_IMAGETAG) VALIDARCHES=amd64 EXCLUDEARCH="$(EXCLUDEARCH)"
 
 ################################################################################
-# Calico cctools image: the CI VM lifecycle binary for the ArgoCI kind-rig.
+# Calico scratch-utils image: the CI VM lifecycle binary for the ArgoCI kind-rig.
 ################################################################################
 
-CALICO_CCTOOLS_IMAGETAG ?= $(shell yq -r '.cctools.version' calico-cctools/versions.yaml)
+CALICO_SCRATCH_UTILS_IMAGETAG ?= $(shell yq -r '.scratch-utils.version' calico-scratch-utils/versions.yaml)
 
 # amd64 only, like binfmt -- the CI/CD runners and the GCE VMs are x86_64.
-.PHONY: calico-cctools-image
-calico-cctools-image: build
-	$(DOCKER_BUILD) -t $(CALICO_CCTOOLS):$(CALICO_CCTOOLS_IMAGETAG)-amd64 -f calico-cctools/Dockerfile calico-cctools/
-	$(MAKE) BUILD_IMAGES=$(CALICO_CCTOOLS) retag-build-images-with-registries VALIDARCHES=amd64 LATEST_IMAGE_TAG=$(CALICO_CCTOOLS_IMAGETAG) IMAGETAG=$(CALICO_CCTOOLS_IMAGETAG)
+.PHONY: calico-scratch-utils-image
+calico-scratch-utils-image: build
+	$(DOCKER_BUILD) -t $(CALICO_SCRATCH_UTILS):$(CALICO_SCRATCH_UTILS_IMAGETAG)-amd64 -f calico-scratch-utils/Dockerfile calico-scratch-utils/
+	$(MAKE) BUILD_IMAGES=$(CALICO_SCRATCH_UTILS) retag-build-images-with-registries VALIDARCHES=amd64 LATEST_IMAGE_TAG=$(CALICO_SCRATCH_UTILS_IMAGETAG) IMAGETAG=$(CALICO_SCRATCH_UTILS_IMAGETAG)
 
-.PHONY: calico-cctools-cd
-calico-cctools-cd: calico-cctools-image var-require-one-of-CONFIRM-DRYRUN
-	$(MAKE) BUILD_IMAGES=$(CALICO_CCTOOLS) retag-build-images-with-registries push-images-to-registries IMAGETAG=$(CALICO_CCTOOLS_IMAGETAG) LATEST_IMAGE_TAG=$(CALICO_CCTOOLS_IMAGETAG) VALIDARCHES=amd64 EXCLUDEARCH="$(EXCLUDEARCH)"
+.PHONY: calico-scratch-utils-cd
+calico-scratch-utils-cd: calico-scratch-utils-image var-require-one-of-CONFIRM-DRYRUN
+	$(MAKE) BUILD_IMAGES=$(CALICO_SCRATCH_UTILS) retag-build-images-with-registries push-images-to-registries IMAGETAG=$(CALICO_SCRATCH_UTILS_IMAGETAG) LATEST_IMAGE_TAG=$(CALICO_SCRATCH_UTILS_IMAGETAG) VALIDARCHES=amd64 EXCLUDEARCH="$(EXCLUDEARCH)"
 
 ################################################################################
 # Calico rust-build image with Rust toolchain.
